+++ src/testing/context.ts
@@ -1,7 +1,5 @@
-import { getLogger } from "@logtape/logtape";
 import type { Context, RequestContext } from "../federation/context.ts";
 import { RouterError } from "../federation/router.ts";
-import { CryptographicKey } from "../vocab/vocab.ts";
 import { mockDocumentLoader } from "./docloader.ts";
 
 export function createContext<TContextData>(
@@ -22,7 +20,6 @@ export function createContext<TContextData>(
     getFeaturedTagsUri,
     parseUri,
     getActorKeyPairs,
-    getActorKey,
     getDocumentLoader,
     sendActivity,
   }: Partial<Context<TContextData>> & { url?: URL; data: TContextData },
@@ -55,28 +52,6 @@ export function createContext<TContextData>(
       throw new Error("Not implemented");
     }),
     getActorKeyPairs: getActorKeyPairs ?? ((_handle) => Promise.resolve([])),
-    getActorKey: getActorKey ?? (async (handle) => {
-      getLogger(["fedify", "federation"]).warn(
-        "Context.getActorKeys() method is deprecated; " +
-          "use Context.getActorKeyPairs() method instead.",
-      );
-      if (getActorKeyPairs == null) return null;
-      for (const keyPair of await getActorKeyPairs(handle)) {
-        const { privateKey } = keyPair;
-        if (
-          privateKey.algorithm.name !== "RSASSA-PKCS1-v1_5" ||
-          (privateKey.algorithm as unknown as { hash: { name: string } }).hash
-              .name !==
-            "SHA-256"
-        ) continue;
-        return new CryptographicKey({
-          id: keyPair.keyId,
-          owner: getActorUri?.(handle) ?? throwRouteError(),
-          publicKey: keyPair.publicKey,
-        });
-      }
-      return null;
-    }),
     sendActivity: sendActivity ?? ((_params) => {
       throw new Error("Not implemented");
     }),